# backend/swagger.json
{
    "openapi": "3.0.3",
    "info": {
      "title": "Project Social Network API",
      "description": "Project Social Network API - by Bootcamp FullStack SysMap",
      "contact": {
        "email": "alexandre.xoquito@gmail.com"
      },
      "version": "1.0.0"
    },
    "servers": [
      {
        "url": "http://localhost:4000/v1"
      }
    ],
...
    "paths": {
...
      "/users/me": {
        "get": {
          "tags": ["users"],
          "summary": "Find logged user",
          "description": "Find logged user",
          "operationId": "findLoggedUser",
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/UserResponse"
                  }
                }
              }
            },
            "401": {
              "description": "UNAUTHORIZED"
            }
          },
          "security": [
            {
              "BearerAuth": {}
            }
          ]
        },
        "put": {
          "tags": ["users"],
          "summary": "Updates logged user's username",
          "description": "Updates username",
          "operationId": "updateUser",
          "requestBody": {
            "description": "Updates logged in username",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserUpdateRequest"
                }
              }
            }
          },
          "responses": {
            "203": {
              "description": "UNDOCUMENTED",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/UserResponse"
                  }
                }
              }
            },
            "401": {
              "description": "UNAUTHORIZED"
            }
          },
          "security": [
            {
              "BearerAuth": {}
            }
          ]
        },
        "delete": {
          "tags": ["users"],
          "summary": "Deletes logged user",
          "description": "deletes logged user",
          "operationId": "deleteUser",
          "responses": {
            "204": {
              "description": "NO CONTENT"
            },
            "401": {
              "description": "UNAUTHORIZED"
            }
          },
          "security": [
            {
              "BearerAuth": {}
            }
          ]
        }
      },
...
    },
    "components": {
      "schemas": {
...
        "UserResponse": {
          "type": "object",
          "properties": {
            "_id": {
              "type": "string",
              "example": "63879d1e2f7daf7fbe0a9f32"
            },
            "user": {
              "type": "string",
              "example": "Lizzera"
            },
            "profile": {
              "type": "string",
              "example": "63879c782f7daf7fbe0a9f2b"
            }
          }
        },
        "UserUpdateRequest": {
          "type": "object",
          "properties": {
            "user": {
              "type": "string",
              "example": "Lizzera"
            },
            "password": {
              "type": "string",
              "example": "q1w2e3"
            }

          }
        },
...
      },
      "securitySchemes": {
        "BearerAuth": {
          "type": "http",
          "scheme": "bearer"
        }
      }
    }
  }